+++ apps/client/src/app/modules/list/+state/lists.selectors.ts
@@ -53,9 +53,15 @@ const getSelectedId = createSelector(
   (state: ListsState) => state.selectedId
 );
 
-const getCurrentListHistory = createSelector(
+const getListHistories = createSelector(
   getListsState,
-  (state: ListsState) => state.selectedListHistory
+  (state: ListsState) => state.listHistories
+);
+
+const getCurrentListHistory = createSelector(
+  getListHistories,
+  getSelectedId,
+  (histories, id) => histories[id]
 );
 
 const getSelectedList = () => createSelector(
@@ -89,5 +95,6 @@ export const listsQuery = {
   getListsLoading,
   getConnectedTeams,
   getSelectedId,
+  getListHistories,
   getCurrentListHistory
 };