+++ apps/client/src/environments/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `376489bb80e467c587a45bae563ed1d629442404`;
+export const extractsHash = `f9833bd87e8b102215d99cb81b89cd88b5815e43`;